+++ lib/rig_kafka/client.ex
@@ -422,7 +422,8 @@ defmodule RigKafka.Client do
 
   # ---
 
-  defp compute_kafka_partition(_topic, n_partitions, key, _value) do
+  defp compute_kafka_partition(_topic, n_partitions, key, _value)
+       when byte_size(key) > 0 do
     partition =
       key
       |> Murmur.hash_x86_32()
@@ -430,5 +431,11 @@ defmodule RigKafka.Client do
       |> rem(n_partitions)
 
     {:ok, partition}
+  end
+
+  defp compute_kafka_partition(_topic, n_partitions, _key, _value) do
+    # based on: https://github.com/klarna/brod#produce-with-random-partitioner
+    random_partition = :crypto.rand_uniform(0, n_partitions)
+    {:ok, random_partition}
   end
 end