+++ test/rig_tests/proxy/publish_to_event_stream/kafka_test.exs
@@ -137,6 +137,239 @@ defmodule RigTests.Proxy.PublishToEventStream.KafkaTest do
                 secured: false,
                 method: "POST",
                 path: endpoint_path,
+                target: "kafka",
+                topic: "rig-test"
+              }
+            ]
+          }
+        },
+        proxy: %{
+          target_url: "http://localhost",
+          port: 3000
+        }
+      })
+
+    headers = [{"content-type", "application/json"}]
+    HTTPoison.post!(rig_api_url, setup_req_body, headers)
+
+    # The client calls the proxy endpoint:
+    request_url = rig_proxy_url <> endpoint_path
+
+    req_body =
+      Jason.encode!(%{
+        "event" => %{
+          "specversion" => "0.2",
+          "type" => "com.example.test",
+          "source" => "/rig-test",
+          "id" => "069711bf-3946-4661-984f-c667657b8d85",
+          "time" => "2018-04-05T17:31:00Z",
+          "data" => %{
+            "foo" => "bar"
+          }
+        },
+        "partition" => "test_key"
+      })
+
+    :timer.sleep(5_000)
+
+    # Send request with set partition key
+    %HTTPoison.Response{status_code: res_status, body: res_body} =
+      HTTPoison.post!(request_url, req_body, headers)
+
+    assert res_status == 202, "Unexpected status #{res_status}: #{res_body}"
+    assert res_body == "Accepted."
+
+    assert_receive received_msg, 10_000
+    received_msg_map = Jason.decode!(received_msg)
+
+    # The event payload is still there:
+    assert get_in(received_msg_map, ["data", "foo"]) == "bar"
+
+    # The event context attributes are, too:
+    assert get_in(received_msg_map, ["type"]) == "com.example.test"
+    assert get_in(received_msg_map, ["id"]) == "069711bf-3946-4661-984f-c667657b8d85"
+
+    # RIG adds meta data to the "rig" extension attribute:
+    assert get_in(received_msg_map, ["rig", "path"]) == "/mock-proxy-publish-to-kafka-endpoint"
+    assert get_in(received_msg_map, ["rig", "remoteip"]) == "127.0.0.1"
+    assert get_in(received_msg_map, ["rig", "correlation"]) |> byte_size > 0
+
+    assert get_in(received_msg_map, ["rig", "headers"])
+           |> Enum.member?(["host", "#{@proxy_host}:#{@proxy_port}"])
+
+    # Send request with no partition key -- RIG will randomize it
+    req_body_random_partition =
+      Jason.encode!(%{
+        "event" => %{
+          "specversion" => "0.2",
+          "type" => "com.example.test",
+          "source" => "/rig-test",
+          "id" => "069711bf-3946-4661-984f-c667657b8d85",
+          "time" => "2018-04-05T17:31:00Z",
+          "data" => %{
+            "foo" => "random partition"
+          }
+        }
+      })
+
+    %HTTPoison.Response{status_code: res_status, body: res_body} =
+      HTTPoison.post!(request_url, req_body_random_partition, headers)
+
+    assert res_status == 202, "Unexpected status #{res_status}: #{res_body}"
+    assert res_body == "Accepted."
+
+    assert_receive received_msg, 10_000
+    received_msg_map = Jason.decode!(received_msg)
+
+    # The event payload is still there:
+    assert get_in(received_msg_map, ["data", "foo"]) == "random partition"
+    assert get_in(received_msg_map, ["rig", "path"]) == "/mock-proxy-publish-to-kafka-endpoint"
+  end
+
+  @tag :avro
+  test "Given target is set to Kafka and schema field is set as well, the http request should publish encoded message to Kafka topic" do
+    test_name = "proxy-publish-to-kafka-avro"
+
+    api_id = "mock-#{test_name}-api"
+    endpoint_id = "mock-#{test_name}-endpoint"
+    endpoint_path = "/#{endpoint_id}"
+
+    # We register the endpoint with the proxy:
+    rig_api_url = "http://localhost:#{@api_port}/v2/apis"
+    rig_proxy_url = "http://localhost:#{@proxy_port}"
+
+    setup_req_body =
+      Jason.encode!(%{
+        id: api_id,
+        name: "Mock API",
+        version_data: %{
+          default: %{
+            endpoints: [
+              %{
+                id: endpoint_id,
+                type: "http",
+                secured: false,
+                method: "POST",
+                path: endpoint_path,
+                target: "kafka",
+                topic: "rig-avro",
+                schema: "rig-avro-value"
+              }
+            ]
+          }
+        },
+        proxy: %{
+          target_url: "http://localhost",
+          port: 3000
+        }
+      })
+
+    headers = [{"content-type", "application/json"}]
+    HTTPoison.post!(rig_api_url, setup_req_body, headers)
+
+    # The client calls the proxy endpoint:
+    request_url = rig_proxy_url <> endpoint_path
+
+    req_body =
+      Jason.encode!(%{
+        "event" => %{
+          "specversion" => "0.2",
+          "type" => "com.example.test",
+          "source" => "/rig-test",
+          "id" => "069711bf-3946-4661-984f-c667657b8d85",
+          "time" => "2018-04-05T17:31:00Z",
+          "data" => %{
+            "foo" => "bar"
+          }
+        },
+        "partition" => "test_key"
+      })
+
+    :timer.sleep(5_000)
+
+    # Send request with set partition key
+    %HTTPoison.Response{status_code: res_status, body: res_body} =
+      HTTPoison.post!(request_url, req_body, headers)
+
+    assert res_status == 202, "Unexpected status #{res_status}: #{res_body}"
+    assert res_body == "Accepted."
+
+    assert_receive received_msg, 10_000
+    received_msg_map = Jason.decode!(received_msg)
+
+    # The event payload is still there:
+    assert get_in(received_msg_map, ["data", "foo"]) == "bar"
+
+    # The event context attributes are, too:
+    assert get_in(received_msg_map, ["type"]) == "com.example.test"
+    assert get_in(received_msg_map, ["id"]) == "069711bf-3946-4661-984f-c667657b8d85"
+
+    # RIG adds meta data to the "rig" extension attribute:
+    assert get_in(received_msg_map, ["rig", "path"]) ==
+             "/mock-proxy-publish-to-kafka-avro-endpoint"
+
+    assert get_in(received_msg_map, ["rig", "remoteip"]) == "127.0.0.1"
+    assert get_in(received_msg_map, ["rig", "correlation"]) |> byte_size > 0
+
+    assert get_in(received_msg_map, ["rig", "headers"])
+           |> Enum.member?(["host", "#{@proxy_host}:#{@proxy_port}"])
+
+    # Send request with no partition key -- RIG will randomize it
+    req_body_random_partition =
+      Jason.encode!(%{
+        "event" => %{
+          "specversion" => "0.2",
+          "type" => "com.example.test",
+          "source" => "/rig-test",
+          "id" => "069711bf-3946-4661-984f-c667657b8d85",
+          "time" => "2018-04-05T17:31:00Z",
+          "data" => %{
+            "foo" => "random partition"
+          }
+        }
+      })
+
+    %HTTPoison.Response{status_code: res_status, body: res_body} =
+      HTTPoison.post!(request_url, req_body_random_partition, headers)
+
+    assert res_status == 202, "Unexpected status #{res_status}: #{res_body}"
+    assert res_body == "Accepted."
+
+    assert_receive received_msg, 10_000
+    received_msg_map = Jason.decode!(received_msg)
+
+    # The event payload is still there:
+    assert get_in(received_msg_map, ["data", "foo"]) == "random partition"
+
+    assert get_in(received_msg_map, ["rig", "path"]) ==
+             "/mock-proxy-publish-to-kafka-avro-endpoint"
+  end
+
+  @tag :kafka
+  test "Given topic and schema fields are not set, the http request should publish message to Kafka topic using fallback method" do
+    test_name = "proxy-publish-to-kafka-fallback"
+
+    api_id = "mock-#{test_name}-api"
+    endpoint_id = "mock-#{test_name}-endpoint"
+    endpoint_path = "/#{endpoint_id}"
+
+    # We register the endpoint with the proxy:
+    rig_api_url = "http://localhost:#{@api_port}/v2/apis"
+    rig_proxy_url = "http://localhost:#{@proxy_port}"
+
+    setup_req_body =
+      Jason.encode!(%{
+        id: api_id,
+        name: "Mock API",
+        version_data: %{
+          default: %{
+            endpoints: [
+              %{
+                id: endpoint_id,
+                type: "http",
+                secured: false,
+                method: "POST",
+                path: endpoint_path,
                 target: "kafka"
               }
             ]
@@ -171,6 +404,7 @@ defmodule RigTests.Proxy.PublishToEventStream.KafkaTest do
 
     :timer.sleep(5_000)
 
+    # Send request with set partition key
     %HTTPoison.Response{status_code: res_status, body: res_body} =
       HTTPoison.post!(request_url, req_body, headers)
 
@@ -188,11 +422,160 @@ defmodule RigTests.Proxy.PublishToEventStream.KafkaTest do
     assert get_in(received_msg_map, ["id"]) == "069711bf-3946-4661-984f-c667657b8d85"
 
     # RIG adds meta data to the "rig" extension attribute:
-    assert get_in(received_msg_map, ["rig", "path"]) == "/mock-proxy-publish-to-kafka-endpoint"
+    assert get_in(received_msg_map, ["rig", "path"]) ==
+             "/mock-proxy-publish-to-kafka-fallback-endpoint"
+
     assert get_in(received_msg_map, ["rig", "remoteip"]) == "127.0.0.1"
     assert get_in(received_msg_map, ["rig", "correlation"]) |> byte_size > 0
 
     assert get_in(received_msg_map, ["rig", "headers"])
            |> Enum.member?(["host", "#{@proxy_host}:#{@proxy_port}"])
+
+    # Send request with no partition key -- RIG will randomize it
+    req_body_random_partition =
+      Jason.encode!(%{
+        "event" => %{
+          "specversion" => "0.2",
+          "type" => "com.example.test",
+          "source" => "/rig-test",
+          "id" => "069711bf-3946-4661-984f-c667657b8d85",
+          "time" => "2018-04-05T17:31:00Z",
+          "data" => %{
+            "foo" => "random partition"
+          }
+        }
+      })
+
+    %HTTPoison.Response{status_code: res_status, body: res_body} =
+      HTTPoison.post!(request_url, req_body_random_partition, headers)
+
+    assert res_status == 202, "Unexpected status #{res_status}: #{res_body}"
+    assert res_body == "Accepted."
+
+    assert_receive received_msg, 10_000
+    received_msg_map = Jason.decode!(received_msg)
+
+    # The event payload is still there:
+    assert get_in(received_msg_map, ["data", "foo"]) == "random partition"
+
+    assert get_in(received_msg_map, ["rig", "path"]) ==
+             "/mock-proxy-publish-to-kafka-fallback-endpoint"
+  end
+
+  @tag :avro
+  test "Given topic and schema fields are not set, the http request should publish encoded message to Kafka topic using fallback method" do
+    test_name = "proxy-publish-to-kafka-avro"
+
+    api_id = "mock-#{test_name}-api"
+    endpoint_id = "mock-#{test_name}-endpoint"
+    endpoint_path = "/#{endpoint_id}"
+
+    # We register the endpoint with the proxy:
+    rig_api_url = "http://localhost:#{@api_port}/v2/apis"
+    rig_proxy_url = "http://localhost:#{@proxy_port}"
+
+    setup_req_body =
+      Jason.encode!(%{
+        id: api_id,
+        name: "Mock API",
+        version_data: %{
+          default: %{
+            endpoints: [
+              %{
+                id: endpoint_id,
+                type: "http",
+                secured: false,
+                method: "POST",
+                path: endpoint_path,
+                target: "kafka"
+              }
+            ]
+          }
+        },
+        proxy: %{
+          target_url: "http://localhost",
+          port: 3000
+        }
+      })
+
+    headers = [{"content-type", "application/json"}]
+    HTTPoison.post!(rig_api_url, setup_req_body, headers)
+
+    # The client calls the proxy endpoint:
+    request_url = rig_proxy_url <> endpoint_path
+
+    req_body =
+      Jason.encode!(%{
+        "event" => %{
+          "specversion" => "0.2",
+          "type" => "com.example.test",
+          "source" => "/rig-test",
+          "id" => "069711bf-3946-4661-984f-c667657b8d85",
+          "time" => "2018-04-05T17:31:00Z",
+          "data" => %{
+            "foo" => "bar"
+          }
+        },
+        "partition" => "test_key"
+      })
+
+    :timer.sleep(5_000)
+
+    # Send request with set partition key
+    %HTTPoison.Response{status_code: res_status, body: res_body} =
+      HTTPoison.post!(request_url, req_body, headers)
+
+    assert res_status == 202, "Unexpected status #{res_status}: #{res_body}"
+    assert res_body == "Accepted."
+
+    assert_receive received_msg, 10_000
+    received_msg_map = Jason.decode!(received_msg)
+
+    # The event payload is still there:
+    assert get_in(received_msg_map, ["data", "foo"]) == "bar"
+
+    # The event context attributes are, too:
+    assert get_in(received_msg_map, ["type"]) == "com.example.test"
+    assert get_in(received_msg_map, ["id"]) == "069711bf-3946-4661-984f-c667657b8d85"
+
+    # RIG adds meta data to the "rig" extension attribute:
+    assert get_in(received_msg_map, ["rig", "path"]) ==
+             "/mock-proxy-publish-to-kafka-avro-endpoint"
+
+    assert get_in(received_msg_map, ["rig", "remoteip"]) == "127.0.0.1"
+    assert get_in(received_msg_map, ["rig", "correlation"]) |> byte_size > 0
+
+    assert get_in(received_msg_map, ["rig", "headers"])
+           |> Enum.member?(["host", "#{@proxy_host}:#{@proxy_port}"])
+
+    # Send request with no partition key -- RIG will randomize it
+    req_body_random_partition =
+      Jason.encode!(%{
+        "event" => %{
+          "specversion" => "0.2",
+          "type" => "com.example.test",
+          "source" => "/rig-test",
+          "id" => "069711bf-3946-4661-984f-c667657b8d85",
+          "time" => "2018-04-05T17:31:00Z",
+          "data" => %{
+            "foo" => "random partition"
+          }
+        }
+      })
+
+    %HTTPoison.Response{status_code: res_status, body: res_body} =
+      HTTPoison.post!(request_url, req_body_random_partition, headers)
+
+    assert res_status == 202, "Unexpected status #{res_status}: #{res_body}"
+    assert res_body == "Accepted."
+
+    assert_receive received_msg, 10_000
+    received_msg_map = Jason.decode!(received_msg)
+
+    # The event payload is still there:
+    assert get_in(received_msg_map, ["data", "foo"]) == "random partition"
+
+    assert get_in(received_msg_map, ["rig", "path"]) ==
+             "/mock-proxy-publish-to-kafka-avro-endpoint"
   end
 end